Transaction in UnitOfWork, ConfigUtils

+++ BaseBackend.Infrastructure/Repository/Base/BaseRepository.cs
@@ -1,13 +1,14 @@
-﻿using BaseBackend.Domain;
+﻿using BaseBackend.CacheManager;
+using BaseBackend.Domain;
 using Dapper;
 using System.Reflection;
-using static Dapper.SqlMapper;
 
 namespace BaseBackend.Infrastructure
 {
     public class BaseRepository
     {
-        public async Task<BaseEntity> InsertItemAsync<TEntity>(BaseEntity entity) where TEntity : BaseEntity
+
+        public (string query, DynamicParameters param, PropertyInfo? idProperty) GetDataForInsert<TEntity>(BaseEntity entity) where TEntity : BaseEntity
         {
             var tableName = entity.GetTableName();
             List<PropertyInfo> _propertyInfo = typeof(TEntity).GetProperties()
@@ -37,51 +38,65 @@ namespace BaseBackend.Infrastructure
             {
                 param.Add($"@{property.Name}", property.GetValue(entity));
             }
+            return (query, param, idProperty);
+        }
+        public async Task<BaseEntity> InsertItemAsync<TEntity>(BaseEntity entity, UnitOfWork unitOfWork) where TEntity : BaseEntity
+        {
+            var data = GetDataForInsert<TEntity>(entity);
             // Thực thi truy vấn
-            using UnitOfWork unitOf = new();
-            int id = await unitOf.Connection.ExecuteScalarAsync<int>(query, param, transaction: unitOf.Transaction);
-            idProperty?.SetValue(entity, id);
+            int id = await unitOfWork.Connection.ExecuteScalarAsync<int>(data.query, data.param, transaction: unitOfWork.Transaction);
+            data.idProperty?.SetValue(entity, id);
             return entity;
         }
 
-        public BaseEntity InsertItem<TEntity>(TEntity entity) where TEntity : BaseEntity
+        public BaseEntity InsertItem<TEntity>(TEntity entity, UnitOfWork unitOfWork) where TEntity : BaseEntity
+        {
+            var data = GetDataForInsert<TEntity>(entity);
+            // Thực thi truy vấn
+            int id = unitOfWork.Connection.ExecuteScalar<int>(data.query, data.param, transaction: unitOfWork.Transaction);
+            data.idProperty?.SetValue(entity, id);
+            return entity;
+        }
+
+        public async Task<int> UpdateItemAsync<TEntity>(TEntity entity, UnitOfWork? unitOfWork = null) where TEntity : BaseEntity
         {
             var tableName = entity.GetTableName();
             List<PropertyInfo> _propertyInfo = typeof(TEntity).GetProperties()
                 .Where(p => p.GetCustomAttributes(typeof(PropertyEntity), false)
                 .Cast<PropertyEntity>()
-                .Any(a => a.IsColumn && !a.IsIdentity)).ToList();
-            List<string> columns = _propertyInfo.Where(p =>
+                .Any(a => a.IsColumn)).ToList();
+            List<string> listUpdateState = [];
+            entity.GetListChangeColumn().ForEach(changeColName =>
             {
-                PropertyEntity? pe = p.GetCustomAttribute<PropertyEntity>();
-                if (pe == null) return false;
-                if (pe.IsColumn && !pe.IsIdentity) return true;
-                else return false;
+                listUpdateState.Add($"{changeColName} = @{changeColName}");
+            });
+            if (listUpdateState.Count > 0 && entity.GetHasVersion())
+            {
+                listUpdateState.Add("version = version + 1");
             }
-            ).Select(p => p.Name).ToList();
-            var values = _propertyInfo.Select(p => $"@{p.Name}").ToList();
-            string columnString = string.Join(", ", columns);
-            string valueString = string.Join(", ", values);
+            string updateState = string.Join(", ", listUpdateState);
+            string query = $"update {tableName} set {updateState} where {entity.GetPrimaryKeyColumnName()} = @{entity.GetPrimaryKeyColumnName()}";
 
-            PropertyInfo? idProperty = typeof(TEntity).GetProperties().Where(p => p.GetCustomAttribute<PropertyEntity>()?.IsIdentity == true).FirstOrDefault();
-            if (idProperty == null)
-            {
-                throw new Exception("Entity must have an identity column.");
-            }
-            string query = $"insert into {tableName} ({columnString}) values ({valueString}) returning {idProperty?.Name};";
             var param = new DynamicParameters();
             foreach (var property in _propertyInfo)
             {
                 param.Add($"@{property.Name}", property.GetValue(entity));
             }
-            // Thực thi truy vấn
-            using UnitOfWork unitOf = new();
-            int id = unitOf.Connection.ExecuteScalar<int>(query, param, transaction: unitOf.Transaction);
-            idProperty?.SetValue(entity, id);
-            return entity;
+            if (unitOfWork != null)
+            {
+                int affectedRows = await unitOfWork.Connection.ExecuteAsync(query, param, transaction: unitOfWork.Transaction);
+                return affectedRows;
+            }
+            else
+            {
+                using UnitOfWork localUnitOfWork = new UnitOfWork();
+                int affectedRows = await localUnitOfWork.Connection.ExecuteAsync(query, param, transaction: localUnitOfWork.Transaction);
+                return affectedRows;
+            }
+
         }
 
-        public async Task<int> UpdateItemAsync<TEntity>(TEntity entity) where TEntity : BaseEntity
+        public int UpdateItem<TEntity>(TEntity entity, UnitOfWork unitOf) where TEntity : BaseEntity
         {
             var tableName = entity.GetTableName();
             List<PropertyInfo> _propertyInfo = typeof(TEntity).GetProperties()
@@ -106,40 +121,8 @@ namespace BaseBackend.Infrastructure
                 param.Add($"@{property.Name}", property.GetValue(entity));
             }
 
-            using UnitOfWork unitOf = new();
-            int affectedRows = await unitOf.Connection.ExecuteAsync(query, param, transaction: unitOf.Transaction);
-
-            return affectedRows;
-        }
-
-        public int UpdateItem<TEntity>(TEntity entity) where TEntity : BaseEntity
-        {
-            var tableName = entity.GetTableName();
-            List<PropertyInfo> _propertyInfo = typeof(TEntity).GetProperties()
-                .Where(p => p.GetCustomAttributes(typeof(PropertyEntity), false)
-                .Cast<PropertyEntity>()
-                .Any(a => a.IsColumn)).ToList();
-            List<string> listUpdateState = [];
-            entity.GetListChangeColumn().ForEach(changeColName =>
-            {
-                listUpdateState.Add($"{changeColName} = @{changeColName}");
-            });
-            if (listUpdateState.Count > 0 && entity.GetHasVersion())
-            {
-                listUpdateState.Add("version = version + 1");
-            }
-            string updateState = string.Join(", ", listUpdateState);
-            string query = $"update {tableName} set {updateState} where {entity.GetPrimaryKeyColumnName()} = @{entity.GetPrimaryKeyColumnName()}";
-
-            var param = new DynamicParameters();
-            foreach (var property in _propertyInfo)
-            {
-                param.Add($"@{property.Name}", property.GetValue(entity));
-            }
-
-            using UnitOfWork unitOf = new();
+            //using UnitOfWork unitOf = new();
             int affectedRows = unitOf.Connection.Execute(query, param, transaction: unitOf.Transaction);
-
             return affectedRows;
         }
         public async Task<int> UpdateItemByColumnName<TEntity>(BaseEntity entity, string colName, string colValue) where TEntity : BaseEntity
@@ -174,16 +157,35 @@ namespace BaseBackend.Infrastructure
             return affectedRows;
         }
 
-        public async Task<int> DeleteItemByColumnName<TEntity>(BaseEntity entity, string colName, string colValue) where TEntity : BaseEntity
+        public async Task<int> DeleteItemByColumnName<TEntity>(BaseEntity entity, string colName, string colValue, UnitOfWork? unitOfWork = null) where TEntity : BaseEntity
         {
             var tableName = entity.GetTableName();
+            string query = "";
+            if (entity.GetHasDeleted())
+            {
+                query = $"update {tableName} set deleted = 1 where {colName} = @where_value";
+            }
+            else
+            {
+                query = $"delete from {tableName} where {colName} = @where_value";
 
-            string query = $"delete from {tableName} where {colName} = @where_value";
-
+            }
             var param = new DynamicParameters();
+            var profile = UserContext.CurrentUser;
             param.Add("@where_value", colValue);
+            param.Add("@updated_by", profile.UserId);
+            param.Add("@updated_time");
             using UnitOfWork unitOf = new();
-            int affectedRows = await unitOf.Connection.ExecuteAsync(query, param, transaction: unitOf.Transaction);
+            int affectedRows = 0;
+            if (unitOfWork != null)
+            {
+                affectedRows = await unitOfWork.Connection.ExecuteAsync(query, param, transaction: unitOfWork.Transaction);
+            }
+            else
+            {
+                using UnitOfWork localUnitOfWork = new UnitOfWork();
+                affectedRows = await localUnitOfWork.Connection.ExecuteAsync(query, param, transaction: localUnitOfWork.Transaction);
+            }
             if (affectedRows == 0) throw new ExecuteErrorException("Delete entity failed!");
             return affectedRows;
         }
